Add nn library
